feat: add test for bug-triggering chunks
We identified a bug in the caching layer that sometimes causes read
to fail. This happen when a read that does not fit into a buffer is
followed by and preceded by reads that do fit. Let's add a test with
this access pattern to make sure the bug is really fixed.